Debug and ensure homo/heteroatomic bonds are properly accounted for.
Atomgraph now contains a new method called getMixing, which takes in chemical ordering and an optional length-2 long int np array (which is modified in place). If no such list is supplied, the method creates and returns one for you. The backend for this is written in C, so libraries have been recompiled in Windows (they must still be compiled for Unix). Interface written for interface.py.

Signed-off-by: AcylSilane <jrd101@pitt.edu>